Dc 124 hotfix add rounding to (re)active power setpoints sent out on FIMS
Original bug: the setpoint is 4.57 which rounds up to 5 and the
component value is 4 (4.57 rounded down) so we spam our command of 4.57

Solution: round the setpoint issued by site controller to its components
but preserve the raw, unrounded value internally. This means we would
send 5 until the component received 5 in the above example.

Adds an optional variable (which I don't like) to the
send_to_components() function for rounding floats so the original value
does not have to be modified

There's an edge case where the value returned on fims gets will not be
rounded. Because only the setpoint component control values should be
rounded, I didn't really have a clean rule I could write for this and I
didn't want to make hardcoded checks against the name to search for
(re)active_power_setpoint. I don't think anybody is using these
endpoints so I just left the issue as is

I wrote a separate approach that rounds the internal value in [this
PR](https://github.com/flexgen-power/hybridos/pull/29). I'm on the fence
about which approach to go with

pytesting