DC-91 PR commit (Asset Templating Revamp) (New way to template with ranges) (#42)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-91)

## Description

Templating revamp based off of integration requests. 

There are now two versions of templating.

### Traditional Templating
This version of templating uses the `number_of_instances` field and
behaves exactly like before.

### Range Based Templating
Allows users to provide a `range` field that allows them to template
non-sequentially.

## How Has This Been Tested?
There are pytest that rely on a specific branch of
[config](https://github.com/flexgen-power/config/tree/AGT_11.3). The
pytest verify that the site starts up and that component pubs go to the
correct asset.

This config has both versions of templating ess is range based and solar
is traditional.

## Quick primer before code review
I am using maps to keep track of adding ids names and component_ids.
There are two types. 
1. The <string, map> all ids/names/component_ids go into one of these so
each template/non_template has a map
a. The inner map is <int, bool>where int will be the asset number for a
template and bool is if it's configured
2. A flat map of all replaced ids/names/component_ids 
a. This contains the finalized ids/names/component_ids and performs
sanity checks that we don't double add anything (aka no collisions)

## Links to updated confluence docu
I created
[this](https://flexgen.atlassian.net/wiki/spaces/API/pages/301858827/Creating+Asset+Templates)
page and linked it in all the Asset configuration pages under the
appropriate field description.