USPOD-935 Fix errant merge (#136)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/USPOD-935)

## Description
Fix HybridOS web server crashing on run, due to a change that set
`loggingService` as an injectable. This PR undoes that change. The
pre-commit linter also now works, which causes the `spec.ts` file to
undergo changes as well.

## How Has This Been Tested?
Build and run hybridos. Observe that no errors appear from the run
script and that the UI loads successfully.