ACO-58-Asset-Manager-Datamaps (#133)
## [Link to JIRA
Ticket](https://flexgen.atlassian.net/jira/software/c/projects/ACO/boards/51?modal=detail&selectedIssue=ACO-58)

## Description
DataMaps are used to run low level/ simulink functions directly from the
ess controller. DataMaps interface with the ess controller's amap to
extract values, run a function with those values, and then update the
amap with the proper output values of the function we just ran.
dataMaps.cpp has been added to /src/ along with a small change to
asset.h to include the definition of the DataMap struct. SL_files folder
has been included to test simulink integration with the ess controller,
and the /scripts/functional_ess/ folder has been added to run the
DataMap function.
The most important part of the code changes are the underlying DataMap
handling functions that extract, store, manipulate, and send data to and
from the amap. The current functions that test and run data maps will be
rearranged when threading is added. Threading will be incorporated in a
later iteration of the code to protect the uses of amap values in other
parts of the ess controller.

## How Has This Been Tested?
The bulk of the testing has been ensuring that the transfer of data to
and from the DataMap and amap has the expected and consistent behavior.
Previous iterations of the code tested this functionality in an isolated
area, and the current iteration shows the flow of data out of the amap,
into the DataMap, and back from the DataMap into the amap. The current
testing function, runDataMaps, will run on the scheduler and show this
functionality. Testing only works on the existing instance of the
DataMap, but more general and configurable set up and run DataMap
functions will be implemented in a later iteration.

---------

Co-authored-by: Preston <pmiller@flexgen.com>
Co-authored-by: Graham Briggs <gbriggs@flexgen.com>
Co-authored-by: pmiller-flexgen <50499111+pmiller-flexgen@users.noreply.github.com>