Dc 107 identify and integrate cpp linter (#125)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-107)

## Description
Adds a script `lint.sh` which runs a clang-tidy check on every changed
file on the branch. "changed" simply means there are uncommitted changes
when compared to your remote git branch.

There are 70+ files changed in this PR, but most of it is autogenerated
with the exception of:
+ site_controller/.clang-format
+ site_controller/.clang-tidy
+ site_controller/docs/linter.md
+ site_controller/.lint.sh

At the moment this is an ad hoc script which can be run to enforce
linting guidelines at a basic level for site controller code. In the
future these extensions are planned:
- A pre-push hook to apply linting, or a github action to run on pending
pull requests.
- A design approval for this to be rolled out to all of our CPP code.
- Any edits requested to the linting rules. If you have ideas for this
check out the documentation for clang-tidy. It can be a bit sparse, but
there are a lot of options for rules to apply, specified in a list on
the first line of `.clang-tidy`.

## How Has This Been Tested?
The build still runs and passes all tests; there shouldn't be any
functionality changes.

---------

Co-authored-by: Eli Handwerk <42653335+ekhand@users.noreply.github.com>
Co-authored-by: vagrant <vagrant@air.mshome.net>
Co-authored-by: mkeyes-flexgen <107417405+mkeyes-flexgen@users.noreply.github.com>