USPOD-910 Support for naked body assets (#135)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/USPOD-910)

## Description
Add support for naked body assets for the site status bar. Naked body
assets are defined here as data with a type of
- `number`,
- `boolean`,
- or `string`,

as opposed to a clothed asset of the Object type, e.g.
```
{
  "value": 12345,
  [remainder of object]
}
```

### A limitation
In clothed assets, the data's unit is provided along with the value
displayed. Naked assets do not have an accompanying unit field, so the
value will appear on the UI as-is.

The same applies to the scalar for a number value, so numbers with a
large number of digits will have them all displayed on the UI. For
example, if fims returns a value of `123.456789000001`, then the UI will
render the exact number `123.456789000001` and take up significant
horizontal space. **Please advise if this should change, perhaps with a
max-length applied.**

## How Has This Been Tested?
- Set up the Site Status Bar per our
[docs](https://flexgen.atlassian.net/wiki/spaces/AP/pages/218562588/Non-Page+Specific+Functionality+-+Site+Controller).
- Amend the `site_status.json` with another `dataSources` entry where
the data is naked. An example (which may or may not work depending on
the `config` that you have loaded into HybridOS):
```
        {
            "uri": "/assets/ess/ess_1",
            "field": "soc",
            "dataType": "number",
            "label": "ESS 1 SOC %"
        }
```
- Ensure the `site_status.json` file is loaded into DBI in the
`hybridos_run.sh` script.
```
load_dbi('ui_config', file_document_renames={'CA_configs.json': 'assets', 'layout_configs.json': 'layout', 'site_status.json': 'site_status'})
```
- Build and run HybridOS. In the UI, observe that the naked body asset
point shows up on the Site Status Bar.