Dc 135 Scheduler signal handler for SIGTERM and SIGINT (#96)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-135)

## Description
Adds a signal handler for SIGTERM and SIGINT to scheduler which will log
the signal when it is received and then terminate the program. The log
is intended to help with debugging unexpected process terminations.

## How Has This Been Tested?
Manually send scheduler SIGTERM and SIGINT and see that the signal is
logged and scheduler terminates.
Manually send scheduler SIGKILL and see that scheduler terminates.