Dc 148 site sequence memory fixes (#165)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-148)

## Description

Refactor sequences maps for clear variable ownership to resolve double
frees and other misc memory issues. This was the same idea as the asset
refactor, but in this case it's just moving to modern containers that
are members of the class, as the variables aren't really being shared.

## How Has This Been Tested?

Existing unit and pytests for sequences passing. 
No more invalid frees/deletes in valgrind

Co-authored-by: Jack Shade <jnshade@ncsu.edu>