Dc 75 fix misleading logs for asset maintenance mode and lockdown logic (#166)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-75)

## Description

Makes the logs and event messages more clear and accurate for
maintenance mode and lockdown mode sets. Adds additional cases to set
handling logic to support the better logging.

## How Has This Been Tested?

Added unit tests for the maintenance mode and lockdown mode set handling
function.
Manually restarted site_controller when maintenance mode and lockdown
mode setpoints were stored in dbi. Saw that no event was posted if an
asset was not in maintenance mode when site_controller restarted. An
event will still be posted if the asset should start back up in
maintenance mode. There is still an event posted and an error logged if
there is an attempt to enter lockdown mode when the asset is not in
maintenance mode.