Dc 146 remove fims and other component info from event messages (#181)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-146)

## Description

Improve the clarity and construction of event messages. I touched up
some additional logging as well but made sure to not change the content
of the logging messages apart from fixing typos and content changes will
come in a future story

## How Has This Been Tested?
Running the system by hand and checking various event messages from the
UI. Unfortunately the events module cannot be queried for specific
endpoints so we don't currently have a way to automate testing afaik.