USPOD-931 Downloadable Events (#151)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/USPOD-931)

## Description
Adds the ability for a user to download the currently filtered events in
the Events page. The implementation downloads all of the events in the
table, as opposed to only the few currently displayed on the UI. For
example, if the UI displays "1-10 of 567", then this feature downloads
all 567 events but otherwise maintains the sort order, filters, and
search query.

Allows downloading of the following formats:
- `.txt` - Raw array of JSON data from the API call.
- `.csv` - File is delimited by `,` for each element and a new line for
each row.
- `.xlsx` - Same appearance as the CSV but in Excel workbook format.
- `.pdf` - Converts the data to a PDF of tables.

## How Has This Been Tested?
- `pnpm install`
- Navigate to the Events page. Click the "Download" button to the right
of the Search bar. Select one of the file formats and download the file.
Open the file and observe that the data mirrors the data on the UI.
- Repeat for the other three file formats.
- Change the filters and re-download the files. Observe that the new
files reflect the filter changes.