DC-176 adjust go.mod module name (#199)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-176)

## Description

Adjust the name of the root module to include `flexgen-power` in the
path. It should have included this in the path to begin with - was
overlooked when implementing the file.

The module path needs to match the path to the repository. The path to
the repository `hybridos` is `github.com/flexgen-power/hybridos`

## How Has This Been Tested?