ACO-85-ESSC-Configuration-files-included-in-compilation (#237)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-85)

## Description
stdcfg files have been added as .h include files. The names of these
files still need to be in the init files.

When running with the "-a" option, when loading configuration files, if
a file name has "std_" in it then the config loader will read it in from
the header file. Otherwise the files will be loaded from DBI.

## How Has This Been Tested?
Several config directories from integration_dev have been shown
successfully working with the "-a" startup option when stdcfg files are
not in DBI.