Dc 184 restore ability to modifiy disabled control values from UI (#255)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-184)

## Description
Restores ability to modify values within all numeric input controls on
configurable assets pages.
- Now, users should be able to identify which fields are currently
active/relating to the current mode by noting the bold and blue
textfields
- All numeric input fields on the configurable assets pages, however,
are still enabled and can be used to send fims messages

Also updates the styling of the controls panel on configurable assets
pages to allow the user to drag the panel to expand/minimize it. This
allows for a denser display of the controls. Also fixes small stylistic
inconvenience where tab content "jumped"/changed formatting when
switching between tabs.

## How Has This Been Tested?
Manually tested in the UI

To QA: Run hybridos with the updated web apps changes, navigate to a
configurable assets page (features, ess, solar, etc.). Note that all
text fields/numeric inputs are enabled and able to be typed in. To see
the updated styling changes, visit Features and switch the mode in the
top dropdown. Notice that relevant certain fields are highlighted with
bold, blue styling.


https://github.com/flexgen-power/hybridos/assets/109531653/f6b32b48-e538-4999-b83b-c8331344b583



[DC-184]:
https://flexgen.atlassian.net/browse/DC-184?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ