DC-255: BUG | Default Value for Numeric Controls is 2 (#253)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-255)

## Description
Fixes small bug in UI where the placeholder value for numeric inputs was
being sent in as 2 (default value for placeholder). Updates it to be
displayed as 0 to avoid confusion

## How Has This Been Tested?
Manually tested in UI

To QA: run UI code, delete a value from a numeric textfield the control
panel of a configurable asset page (ESS, PCS, Features, etc.). See that
once the value is deleted it is replaced by a 0, rather than a 2

[DC-255]:
https://flexgen.atlassian.net/browse/DC-255?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Kyle Brezina <50492569+kylebrezina@users.noreply.github.com>