DC-260: Scheduler Events Scheduled from UI in Different Time Zone Use the Browsers Time Zone (#267)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-260)

## Description
Updates code to take into account the timezone coming back from
/scheduler/timezones URI when scheduling event. Previously, the datetime
library was using the timezone pulled from the browser time. Now, all
times/dates shown on the Scheduler UI should be assumed to be in the
timezone for which the individual site is configured.

(i.e. if a site is configured for timezone America/Los_Angeles, events
for that site will be scheduled in that timezone, even if they are being
scheduled from a computer which is is the Eastern timezone)

## How Has This Been Tested?
Manually QA'd in UI

Steps to QA: 
- In the vm, update your system time to be in a different timezone than
your local time by running
`sudo timedatectl set-timezone America/Los_Angeles`
- Run the UI and schedule an event on the Scheduler page
- Ensure that when you click on the event, the start/end date and time
look identical to what you scheduled


[DC-260]:
https://flexgen.atlassian.net/browse/DC-260?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ