Functional ESS Controller - InputHandler Functions (BMS and PCS) (#299)
## [Link to JIRA Story](https://flexgen.atlassian.net/browse/ACO-118)
## Description
This pull request creates the InputHandler namespace which holds all of
the functions that can be called from configs for the purpose of
StartBMS, StopBMS, StartPCS, StopPCS, StandbyPCS. These actions would
usually be done from ess_controller within configs, but now much of the
configs have been moved into code. The main portion that remains in
configs is simply putting the functions onto the scheduler.

There is also a new namespace called "HandlerUtility" within the
"IOHandlerUtility.cpp" file which has functions that reduce the amount
of repeated code and instead allow for cleaner and reusable functions.

The InputHandler functions are not useful without their OutputHandler
counterpart but this pull request is to get the initial structure of
InputHandler into dev without having to have a massive change with
everything.


## How Has This Been Tested?
Using the script files within ess_controller. If you go to
/ess_controller/scripts/functional_ess_controller there are two script
files which are "bms.sh" and "pcs.sh". These scripts set up all the uris
necessary to run the InputHandler functions from the scheduler in the
most basic way. These scripts have been run and confirmed to start all
of the InputHandler functions.