Add doxygen examples and documentation on how to use (#306)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-318)

## Description
The attached ticket goes well beyond the contents of this pull request.
The main outcome was the creation of this document:
https://flexgen.atlassian.net/wiki/spaces/FPS/pages/397410346

However the end of that document chats about `doxygen`, a tool which is
added in with this pull request. The consensus at the moment is to not
put a ton of focus on refining this, but it's useful and this small
addition will add value.

## How Has This Been Tested?
To utilize doxygen simply install according to the steps and run the
steps included. You should be able to fire up a documentation server and
navigate the files easily, even for files you had just written.

The files should all be gitignored.