Dc 48 Fix slewing for competing assets under discharge (#289)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-48)

## Description
Active Power Setpoint calculates it's target as the aggregate of the
asset commands (excluding the POI). When discharging or charging from
the POI this works fine, but when the ESS charges off of solar or
generator we'll get an inaccurate target. This means that the charging
command can get stuck at the wrong value

For example, the user enters a command of -10MW indicating they want to
charge 10MW. 10MW of solar is available to handle this charge. That
produces a slew target of 10MW (solar) - 10MW (ESS) = 0MW. So the next
iteration the user's -10MW is still stuck slewing from 0MW rather than
gradually decreasing down to the intended -10MW

## How Has This Been Tested?
- unit tests
- pytests
- manual testing following steps similar to described above.