ESSC BMS contactor control (#308)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-107)

## Description

This pull request has the changes for having full BMS contactor control
with minimal configs needed. Much of the contactor control has been
moved into the code of ess_controller. This is for the purpose of
simplifying what is needed for ess_controller configs.

## How Has This Been Tested?
This has been tested using go_metrics and the script files
"start_bms.sh" and "stop_bms.sh" to be able to confirm that contactor
control is possible with the new ess_controller code.

---------

Co-authored-by: pmiller-flexgen <pmiller@flexgen.com>