Update all sandbox configs script (#322)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-335)

## Description
All Files changed were simply a result of
`sandbox/dispatch_coordination/scripts/update_all.sh.` That should be
the only file that should need scrutinous review. That being said, its
intent is simply to save us time, so it doesn't need to be efficient nor
beautiful, just get the job done of autoapplying updates.

Implement update all script and run over all the files.

The intent is that the updater outputs in a consistent format which
isn't matched by the configs inside of monorepo today. So, if we do one
big run of updater without any changes today, they all enter a format
that will be sensibly diffable in the future. The precise format is for
them to have all keys in alphabetical order.

This way if one of us has a config change needed for their PR, the
intent is to:
- implement the rules inside of updater as the fake 11.3 => 11.4 version
- run update_all.sh
- inspect and commit the changes inside of monorepo.

Brief documentation:
https://flexgen.atlassian.net/wiki/spaces/FPS/pages/427196550/Updating+all+sandbox+configs+unit+test+configs+at+once

## How Has This Been Tested?
Ran over files, I have ran some tests and there should be no changes
whatsoever, barring some flagrant error with the updater which I find
unlikely in this case.
- [x] unit tests
- [x] pytests to some degree, I haven't ran 100% of them. I haven't seen
any evidence of configs changing after the running of this script,
meaning there is no reason to think that pytests would suddenly fail.

---------

Co-authored-by: pmiller-flexgen <50499111+pmiller-flexgen@users.noreply.github.com>
Co-authored-by: Graham Briggs <gbriggs@flexgen.com>
Co-authored-by: John Lawler <84024987+jplawler@users.noreply.github.com>
Co-authored-by: Annie Forgette <109531653+anniekforgette@users.noreply.github.com>
Co-authored-by: Judsen Hembree <105236099+JudsenAtFlexgen@users.noreply.github.com>
Co-authored-by: Emily Applegate <eapplegate@flexgen.com>