DC-267 Publish hardware statistics on /cops/stats/system (#342)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-267)

## Description
Implement the backend of overwatch's collection engine to publish
hardware statistics for the system status page to read/report from via
FIMS.

Specific URIs being published: 
```
/cops/stats/system/net
/cops/stats/system/mem
/cops/stats/system/cpu
/cops/stats/system/disk
```
This involves data specific to the hardware controller itself and
overall usage. This is **not** publishing service specific data such as
CPU and MEM utilization per process.

### Enabling system statistics reporting: 
- there are two methods to enable the status reports: 
     - cmd line flag 
     - configuration

When executing, add `-syswatch` to the execution command to enable.
`/usr/local/bin/cops -c=/config/path -syswatch`
For config, add the following to enable as a top level object in the
COPS configuration file:
```
"syswatch": true,
```

** Side Note:
The COPS go.mod and go.sum file are removed in this story in place of
using the monomodule. Since this story resulted in creating a new
package in COPS, it was easy to fit in removing the COPS submodule in
place of using the monomodule go.mod file instead.


## How Has This Been Tested?
TESTS TODO:
- Implement unit tests for respective packaged logic
- Test against running hybridos using sandbox configs

---------

Co-authored-by: Kyle Brezina <50492569+kylebrezina@users.noreply.github.com>