DC-344 Respond to FIMS with action result (#348)
## [DC-344 FIMS error reply for systemd
actions](https://flexgen.atlassian.net/browse/DC-344)

## Description
Reply to a set on a specific action control with the result of the
action once completed. The result is a single string response that can
be several types: done, canceled, timeout, failed, dependency, skipped
> done indicates successful execution of a job. canceled indicates that
a job
has been canceled before it finished execution. timeout indicates that
the
job timeout was reached. failed indicates that the job failed.
dependency
indicates that a job this job has been depending on failed and the job
hence
has been removed too. skipped indicates that a job was skipped because
it
 didn't apply to the units current state.

## How Has This Been Tested?
- Spin up hybridos instance, enact action via FIMS set and verify
response is as appropriate.