DCM-82 ftd archive handling full archives (#345)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DCM-82)

## Description

Previously when FTD received a message for an encoder that was already
full, it would still try to add the message to the full encoder. This PR
changes the behavior so that whenever any encoder reaches the maximum
message count, FTD will create archives out of all its existing
encoders. That way, we make room to add new messages.

Within fims_codec, trying to encode a message with an encoder that is
already full will now return an error.

## How Has This Been Tested?

All fims_codec unit tests passed.

Ran hybridos with powercloud and ran a script which would periodically
publish a counter at a very fast rate, such that the max message count
was reached. Saw that all data arrived in Influx as expected.

---------

Co-authored-by: pmiller-flexgen <50499111+pmiller-flexgen@users.noreply.github.com>
Co-authored-by: Graham Briggs <gbriggs@flexgen.com>
Co-authored-by: John Lawler <84024987+jplawler@users.noreply.github.com>
Co-authored-by: Annie Forgette <109531653+anniekforgette@users.noreply.github.com>
Co-authored-by: James White <43786415+jw-develop@users.noreply.github.com>
Co-authored-by: Judsen Hembree <105236099+JudsenAtFlexgen@users.noreply.github.com>
Co-authored-by: Emily Applegate <eapplegate@flexgen.com>