DC-367 | UI | Ensure All Pages Appear Correctly for Observer Role (#373)
## [DC-367 | UI | Ensure All Pages Appear Correctly for Observer
Role](https://flexgen.atlassian.net/browse/DC-367)

## Description
Fixes bug where users with Observer roles were able to use some of the
controls on configurable assets pages. The Observer role is a read-only
role and should not have access to any controls.

## How Has This Been Tested?
Manually tested in the UI

To QA:
1. Build the web_apps folder on this branch
2. Run HybridOS, log in with an admin user
3. Navigate to the User Administration tab, and create a new user with
an Observer role
4. Log out of HOS and back in as the new role you just created
5. Click through all the pages of HOS and ensure that you are not able
to interact with any of the controls. You should only be able to see
data, all input fields/buttons should be disabled.