Dc 357 site sequences bugfix (#368)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-357)

## Description

Sets to /events were being spammed due to a refactor to how sequences
tracks alerts. This resulted in the events service becoming
unresponsive.

Also found a bug where incomplete variables configuration could cause
segfaults for certain active_alarm/active_fault configurations.

Head is James's DC-271 as it has a pytest framework change that I didn't
want to duplicate. So only 2 new commits to review.

@kylebrezina first commit is the fix for out of bounds segfaults due to
incomplete configuration. That's the bug that has existed for a while
that could potentially be backported if needed (but probably not).

The second commit is patch for spamming events which is unique to 11.3

## How Has This Been Tested?
unit and pytests