Fix bug with initializing standalone power features & watchdog_enable… (#388)
… (#371)

## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-356)

## Description
Fixes a bug with initializing features as false even if their enable
flag is configured out of the gate with a value of true. This looks to
have been introduced during features encapsulation, in which much of
this code was refactored, and there was a function which initialized
everything as false expecting them to be activated later. The init to
false portion was kept around but the "flip them to true" section was
deleted at some point.

The solution is to get rid of the "falsify all feature enable flags"
section, because they don't need to be there anymore, at least for
standalone and site_operation features which is what the scope of this
ticket is limited to.

There is a small additional configuration enforcement described below.

## How Has This Been Tested?
- [x] new pytest `enable_flags.py`
- [x] unit tests. There were a handful of unit_test_configs which needed
small edits to not violate the "not available but enabled" rule.

## Changes to configuration?
No, though there new error enforcement in the event of this bad
configuration:
- If a feature is not available (via its available_features register),
and it is set to enabled == true, site controller will error out. It's
unlikely this will cause issues with customers; at most they will have
to go in and flip a few enable flags to false as they should be, which
has been deemed reasonable.

---------

## [Link to JIRA Ticket](replacemewithlinktojiraticket)

## Description

## How Has This Been Tested?

---------

Co-authored-by: andrewkwonFlexGen <83837917+andrewkwonFlexGen@users.noreply.github.com>
Co-authored-by: pmiller-flexgen <50499111+pmiller-flexgen@users.noreply.github.com>
Co-authored-by: Graham Briggs <gbriggs@flexgen.com>
Co-authored-by: John Lawler <84024987+jplawler@users.noreply.github.com>
Co-authored-by: Annie Forgette <109531653+anniekforgette@users.noreply.github.com>
Co-authored-by: Judsen Hembree <105236099+JudsenAtFlexgen@users.noreply.github.com>
Co-authored-by: Emily Applegate <eapplegate@flexgen.com>
Co-authored-by: Eli Handwerk <42653335+ekhand@users.noreply.github.com>
Co-authored-by: Philip Wilshire <63882404+philflex2020@users.noreply.github.com>
Co-authored-by: Stephanie Reynolds <sreynolds@flexgen.com>
Co-authored-by: BrandonKnoble1 <105602159+BrandonKnoble1@users.noreply.github.com>