Aco 90 psm leaf assets should ignore child asset expressions (#427)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-90)

## Description
This bugfix prevents PSM from crashing or failing to launch if an
expression in the config flow fails to parse. Instead, the default value
from `psm_dflt.json` will be used in these cases.

## How Has This Been Tested?

Launch PSM with feeder asset as a leaf node. This will cause expressions
for the feeder `pmax` `smax` and `qmax` to fail to parse as these
expressions all look for a child asset which will not exist. Before this
change, this would cause psm to panic. Following this change, psm
launches correctly and uses the default vaules.