Dc 363 create coordinate configurable asset pages (#413)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-363)

## Description

Create configurable assets pages for the reference configs. This was
extended to include config_fleetman_dev as well, and various
scripts/config updates to get that test bench running again.

Still need some UI code changes to get the controls working with the
naked pubs sent by default over dnp3. Also the twins model is still
broken in the sandbox. I don't think twins is even running. But it felt
like this ticket already went way beyond scope so I haven't resolved it
in this ticket

Honestly a lot of config changes that maybe don't need to be reviewed.
The main points for review are:
* I manually set up the network because the automated solution wasn't
working for me
* I reverted the psm rename. I can undo this if/when psm is functional.
I'm a little unclear on the timeline here but just needed something
functional to demo against.
* New scripts updates in hybridos_run.py and vsaap.py

## How Has This Been Tested?
Running the test bench and observing the web pages:
for reference configs cd to the vfakeco directory and `docker compose up
-d` with the fleet and site containers at a minimum.

for the sandbox the run script should now be functional so:
`hybridos_run --init --sandbox config_fleetman_dev`