DC-386 | UI | Web Server | Add Support for Naked Control Publishes for Configurable Assets Pages (#419)
## [DC-386 | UI | Web Server | Add Support for Naked Control Publishes
for Configurable Assets
Pages](https://flexgen.atlassian.net/browse/DC-386)

## Description
Adds support to the web server to receive and correctly display controls
on the Configurable Assets Pages based on fully naked body publishes.
This removes the requirement for control publishes to include a enabled
and value field. If no enabled field is received, this PR updates the
code to assume these controls should be enabled and to use the value
received as the value for these controls.

## How Has This Been Tested?
Manually tested in the UI. The dispatch_coordination sandbox was used to
ensure that existing functionality was retained. The virtual site fleet
manager environment (updated in
[PR-413](https://github.com/flexgen-power/hybridos/pull/413)) was used
to test the new functionality