DC-382 | UI | Web-Server | Add Support for Parsing Maintenance Actions from Configuration File (#416)
## [DC-382 | UI | Web-Server | Add Support for Parsing Maintenance
Actions from Configuration
File](https://flexgen.atlassian.net/browse/DC-382)

## Description
Updates the web server parsing of the ui_config/assets.json
configuration file to accept a new flag, `hasMaintenanceActions`. This
is an optional flag that will be passed to the frontend to determine
whether or not to display maintenance actions on a configurable asset
page.

If the flag is set to true, this PR also sets up the structure to handle
the control to start a new maintenace action as well as incoming pubs
regarding the mainteance actions. Information gathered from FIMS
publishes regarding the mainteance actions are passed to the frontend
within a separate object, with the key maintenanceActions, that falls
within the existing data structure. An example of the updated data
object passed to the frontend is seen below:

## How Has This Been Tested?
Manually tested in the UI