PSM - transformer Vn default should allow for a feeder as a parent (#437)
## [ACO-167](https://flexgen.atlassian.net/browse/ACO-167)

## Description
The transformer model has a nominal voltage parameter `Vn` that needs to
be configured at startup with the an appropriate default (usually the
voltage of the parent asset). In cases where a feeder asset was the
parent of a transformer, then transformer had no parameter to look at as
the feeder asset does not determine its voltages at config time. The fix
is to configure the feeder parent side voltage in the service of
configuring the transformer `Vn`, even though the feeder parent voltage
will be immediately overwritten during run time.
## How Has This Been Tested?
Launch psm with the default config scheme with an asset tree that
includes feeder -> transformer. See that the assets configure correctly
with no expression parsing warnings. Configuring the transformer with n
> 1 will result in the appropriate voltages cascading through the tree.


[ACO-167]:
https://flexgen.atlassian.net/browse/ACO-167?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ