DC-379 | UI | Include data from multiple sources in a single configurable assets entry (#433)
## [DC-379](https://flexgen.atlassian.net/browse/DC-379)

## Description
Adds support for displaying data from multiple source URIs on a single
configurable assets tab/entry. For example, the statuses and controls
for a site in Coordinate can now be sourced from different URIs but
still be displayed on the same tab. Also adds support for URIs to have
duplicate final fragments as well as start with different fragments
(previously using URIs like `/sites/lima` and
`/fleet/features/ercotAs/lima` at the same time was not supported).

## How Has This Been Tested?
Manually tested in the UI.


---------

Co-authored-by: Annie Forgette <aforgette@flexgen.com>
Co-authored-by: Kyle Brezina <50492569+kylebrezina@users.noreply.github.com>