DC-412 Add Cipher suites and set min TLS version in scheduler to dev (#450)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-412)

## Description
When a WebSocket client connects to the Scheduler WebSocket server, the
handshake involves the server having a list of cipher suites that the
client can use to encrypt its messages. Scheduler uses a default set of
cipher suites provided by the tls package. Some of these ciphers are
being marked in a security audit as being too weak.

## How Has This Been Tested?
- Validate Scheduler’s WebSocket port is not called out by Nessus.
- Validate whether or not port is supporting 3DES ciphers we are looking
to remove:
 - `openssl s_client -cipher 3DES -connect localhost:9000` 
- add `-tls1` to check if tls 1.0 is enabled, `-tls1_1` for 1.1. [more
info here
](https://www.openssl.org/docs/man1.0.2/man1/openssl-s_client.html#:~:text=ssl2%2C%20%2Dssl3%2C%20%2Dtls1%2C%20%2Dtls1_1%2C%20%2Dtls1_2%2C%20%2Dno_ssl2%2C%20%2Dno_ssl3%2C%20%2Dno_tls1%2C%20%2Dno_tls1_1%2C%20%2Dno_tls1_2)